backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
numpy
pandas
scikit-learn
tensorflow
joblib
